Add socketManager
